Set encoding to utf8, use cabal 3.10.3.0
This allows us to run inside nix sandbox, need to use an older cabal
as the newer cabal version errors a lot with withFile: resourceBusy
when we provide CABAL_DIR for the project from haskell.nix. It is also
expensive to the cabal version we use from haskell.nix right now.